ref: first pass at refactoring
